requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.4
aiokafka==0.10.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.0
pytest==8.3.3
pytest-asyncio==0.24.0
